refactor: setqueryParams allow removing params, move to url utils
