+++ src/shared/http-utils.spec.ts
@@ -2,7 +2,6 @@ import { EndpointError } from './errors.js';
 import {
   queryXmlDocument,
   setFetchOptions,
-  setQueryParams,
   sharedFetch,
 } from './http-utils.js';
 import { fetchDocument } from '../ogc-api/link-utils.js';
@@ -169,68 +168,6 @@ describe('HTTP utils', () => {
       it('only fetches the document once', async () => {
         expect(globalThis.fetch).toHaveBeenCalledTimes(1);
       });
-    });
-  });
-
-  describe('setQueryParams', () => {
-    it('adds new parameters if not present', () => {
-      expect(
-        setQueryParams('https://my.host/service?arg1=123', {
-          ARG2: '45',
-          Arg3: 'hello',
-        })
-      ).toBe('https://my.host/service?arg1=123&ARG2=45&Arg3=hello');
-    });
-    it('replaces existing parameters regardless of case', () => {
-      expect(
-        setQueryParams('https://my.host/service?ARG1=123&Arg2=bla&arg3', {
-          ARG2: '45',
-          Arg3: 'hello',
-        })
-      ).toBe('https://my.host/service?ARG1=123&ARG2=45&Arg3=hello');
-    });
-    it('sets a parameter without value if true', () => {
-      expect(
-        setQueryParams('https://my.host/service', {
-          ARG2: true,
-        })
-      ).toBe('https://my.host/service?ARG2=');
-    });
-    it('appends an encoded URL if found (HTTP)', () => {
-      expect(
-        setQueryParams('http://bad.proxy/?url=http%3A%2F%2Fmy.host%2Fservice', {
-          ARG2: '45',
-          Arg3: 'hello',
-        })
-      ).toBe(
-        'http://bad.proxy/?url=http%3A%2F%2Fmy.host%2Fservice%3FARG2%3D45%26Arg3%3Dhello'
-      );
-    });
-    it('appends an encoded URL if found (HTTPS)', () => {
-      expect(
-        setQueryParams(
-          'http://bad.proxy/?url=https%3A%2F%2Fmy.host%2Fservice',
-          {
-            ARG2: '45',
-            Arg3: 'hello',
-          }
-        )
-      ).toBe(
-        'http://bad.proxy/?url=https%3A%2F%2Fmy.host%2Fservice%3FARG2%3D45%26Arg3%3Dhello'
-      );
-    });
-    it('makes sure that spaces are encoded as %20', () => {
-      expect(
-        setQueryParams(
-          'https://my.host/service?arg1=old+value&something=else+entirely',
-          {
-            ARG1: 'new value',
-            'ARG 2': 'value with space',
-          }
-        )
-      ).toBe(
-        'https://my.host/service?something=else%20entirely&ARG1=new%20value&ARG%202=value%20with%20space'
-      );
     });
   });
